chore: move AppsBrowser components to own folder
