+++ src/popup/components/AppsBrowser/AppsBrowserHeader.vue
